stage-7: minors
